finish up changing the tab service, just need to test custom tabs and multiple chief complaints
